sensor/requirements.txt
paho-mqtt==2.1.0
smbus==1.1.post2